+++ src/frontend/helper/ScrollToTop.js
@@ -1,0 +1,12 @@
+import { useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
+
+export const ScrollToTop = () => {
+  const { pathname } = useLocation();
+  useEffect(() => {
+    setTimeout(() => {
+      window.scrollTo({ top: 0 });
+    }, 0);
+  }, [pathname]);
+  return null;
+};
