+++ src/frontend/helper/index.js
@@ -1,2 +1,3 @@
 export * from './hooks';
 export * from './PageTemplate';
+export * from './ScrollToTop';
