+++ src/frontend/pages/individualPost/index.js
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom';
 import '../homepage/post.css';
 import { posts } from '../../utility/constants';
 import Post from './Post';
-import { PageTemplate } from '../../helper/PageTemplate';
+import { PageTemplate, ScrollToTop } from '../../helper';
 
 export default function IndividualPost() {
   const { postId } = useParams();
@@ -16,6 +16,7 @@ export default function IndividualPost() {
 
   return (
     <Fragment>
+      <ScrollToTop />
       <PageTemplate>
         <Post post={post} />
       </PageTemplate>
